Add forced floating ip disassociation to script
This addresses a problem where when you associate a floatingip that already has an existing association, warning messages are printed in the q-l3 script.  I've run this twice - make sure it didn't generate its own problems as I try to disassociate a floating ip whether it has an association or not.